+++ R/RcppExports.R
@@ -9,7 +9,11 @@ denseSLOPE <- function(x, y, control) {
     .Call(`_SLOPE_denseSLOPE`, x, y, control)
 }
 
-sorted_l1_prox <- function(x, lambda) {
-    .Call(`_SLOPE_sorted_l1_prox`, x, lambda)
+lambdaSequence <- function(n_lambda, q, theta1, theta2, lambda_type, n) {
+    .Call(`_SLOPE_lambdaSequence`, n_lambda, q, theta1, theta2, lambda_type, n)
+}
+
+sortedL1ProxCpp <- function(x, lambda, method) {
+    .Call(`_SLOPE_sortedL1ProxCpp`, x, lambda, method)
 }