+++ src/RcppExports.cpp
@@ -37,15 +37,32 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// sorted_l1_prox
-arma::mat sorted_l1_prox(const arma::mat& x, const arma::vec& lambda);
-RcppExport SEXP _SLOPE_sorted_l1_prox(SEXP xSEXP, SEXP lambdaSEXP) {
+// lambdaSequence
+arma::vec lambdaSequence(const arma::uword n_lambda, const double q, const double theta1, const double theta2, const std::string lambda_type, const arma::uword n);
+RcppExport SEXP _SLOPE_lambdaSequence(SEXP n_lambdaSEXP, SEXP qSEXP, SEXP theta1SEXP, SEXP theta2SEXP, SEXP lambda_typeSEXP, SEXP nSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const arma::uword >::type n_lambda(n_lambdaSEXP);
+    Rcpp::traits::input_parameter< const double >::type q(qSEXP);
+    Rcpp::traits::input_parameter< const double >::type theta1(theta1SEXP);
+    Rcpp::traits::input_parameter< const double >::type theta2(theta2SEXP);
+    Rcpp::traits::input_parameter< const std::string >::type lambda_type(lambda_typeSEXP);
+    Rcpp::traits::input_parameter< const arma::uword >::type n(nSEXP);
+    rcpp_result_gen = Rcpp::wrap(lambdaSequence(n_lambda, q, theta1, theta2, lambda_type, n));
+    return rcpp_result_gen;
+END_RCPP
+}
+// sortedL1ProxCpp
+arma::mat sortedL1ProxCpp(const arma::mat& x, const arma::vec& lambda, const int method);
+RcppExport SEXP _SLOPE_sortedL1ProxCpp(SEXP xSEXP, SEXP lambdaSEXP, SEXP methodSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< const arma::mat& >::type x(xSEXP);
     Rcpp::traits::input_parameter< const arma::vec& >::type lambda(lambdaSEXP);
-    rcpp_result_gen = Rcpp::wrap(sorted_l1_prox(x, lambda));
+    Rcpp::traits::input_parameter< const int >::type method(methodSEXP);
+    rcpp_result_gen = Rcpp::wrap(sortedL1ProxCpp(x, lambda, method));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -53,7 +70,8 @@ END_RCPP
 static const R_CallMethodDef CallEntries[] = {
     {"_SLOPE_sparseSLOPE", (DL_FUNC) &_SLOPE_sparseSLOPE, 3},
     {"_SLOPE_denseSLOPE", (DL_FUNC) &_SLOPE_denseSLOPE, 3},
-    {"_SLOPE_sorted_l1_prox", (DL_FUNC) &_SLOPE_sorted_l1_prox, 2},
+    {"_SLOPE_lambdaSequence", (DL_FUNC) &_SLOPE_lambdaSequence, 6},
+    {"_SLOPE_sortedL1ProxCpp", (DL_FUNC) &_SLOPE_sortedL1ProxCpp, 3},
     {NULL, NULL, 0}
 };